requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
langchain-community
transformers
torch

